ci(lint): CI has never linted a line of test code
`cargo clippy --workspace --exclude rscx` without `--all-targets` skips every
test, bench and example target in the workspace. 92 warnings had accumulated
there unseen: 84 at default features, 7 more behind `-p scx-accel --features
gpu`, and 1 behind `-p scx-convert --features hdf5` — a lane no CI job compiled
at all, since the `test` job runs default features only.

Adds `--all-targets` to the `clippy` job and to the six non-`--bin` lanes of the
`feature-check` matrix (the `gpu_bench` lane keeps `--bin`, which `--all-targets`
contradicts), adds the missing `scx-convert --features hdf5` lane, and clears
all 92.

About half were machine-applicable (`clippy --fix`). Four of the manual ones
were worth more than the lint:

- Five GPU-vs-CPU parity loops in `gpu_preprocess_tests.rs` and one in
  `neighbors/cpu.rs` indexed `0..cpu_data.len()`. Rewriting them as `zip`
  would have silently truncated the comparison to the shorter side, so each
  gained an explicit length assertion first. Three of the six had no length
  check at all — a short GPU result would have passed.
- `cpu_decode` in `shard_decode.rs` took eight arguments, three of them
  adjacent `&[u8]`. It now takes the `EncodedShardRef` they already form: a
  positional mix-up between two byte-stream parameters is precisely what a
  parity test cannot detect.

No production behaviour changes. `cargo test --workspace --exclude rscx`:
3003 passed, 0 failed. `scx-accel --features gpu`, `scx-gpu` and
`scx-convert --features hdf5` suites all green.

+++ scx-gpu/src/gpu_csr_assemble.rs
@@ -230,7 +230,7 @@ mod tests {
             n_rows,
             nnz,
             header.index_dtype == 0,
-            scx_codec::clamp_index_bound(n_cols as u32),
+            scx_codec::clamp_index_bound(n_cols),
         )
         .unwrap()
     }
